ft.h ok, ft_print_memory stable ! Diverses corrections et mises à la norme mineures sur quelques fonctions.

+++ day05/ex23/ft_print_memory.c
@@ -1,6 +1,6 @@
 void	ft_putchar(char c);
 
-void	putnbr_hexa(int c)
+void	putnbr_hexa(unsigned long int c)
 {
 	char	*base;
